refactor: moved load on params
